[bsg_mem] fix a small bug in mask_write_bit synth model (see more)
Fix for a very small bug in bsg_mem_1rw_sync_mask_write_bit_synth.

To implement the bit write mask, the code was looping over els_p
(elements) instead of over width_p (bits). So it's an easy fix. With
the bug, if els_p < width_p, the upper bits will not be written.