Added XILINX attributes to eliminate FPU timing loops/warnings (#134)
These added attributes solve warnings in the Xilinx tool chain related to:

* Unused registers (Trimmed)
* Resource Sharing

The latter prevents a timing loop from being inferred in the toolchain, or at least prevents the messages from occuring. It is not yet determined if these fixes solve the floating point computation errors seen in Huwan's code in October.